requirements.txt
Flask==2.3.3
PyJWT==2.8.0
gunicorn==20.1.2
